More endpoints

+++ app/api/countries/[slug]/route.ts
@@ -2,7 +2,6 @@ import { db } from "@/db"
 import { CountryInsert, countries } from "@/db/schema"
 import { prepareForPatch } from "@/utils/patch"
 import { eq } from "drizzle-orm"
-import { NextResponse } from "next/server"
 
 export async function GET(
     request: Request,
@@ -31,30 +30,15 @@ export async function DELETE(
     { params }: { params: { slug: number } }
 ) {
     const countryId = params.slug
+    const country = (await db.select().from(countries).where(eq(countries.id, countryId)))[0]
+
+    if (country == null) {
+        return Response.json({ message: "Not found" }, { status: 404 })
+    }
 
     await db.delete(countries).where(eq(countries.id, countryId))
 
     return new Response(null, { status: 204 })
-} 
-
-function validateInput(insertModel: any, updateModel: any) {
-    if (insertModel.id != null) { 
-        return Response.json({ message: "ID must be null" }, { status: 400 }) 
-    }
-
-    for (const prop in insertModel) {
-        if (!(prop in updateModel)) {
-            return Response.json({ message: `Property '${prop}' is not valid` }, { status: 400 })
-        }
-    }
-
-    for (const prop in updateModel) {
-        if (!(prop in insertModel)) {
-            return Response.json({ message: `Property '${prop}' in is not val id` }, { status: 400 });
-        }
-    }
-
-    return null
 }
 
 export const dynamic = 'force-dynamic'
